requirements.txt
albumentations==2.0.8
clearml==2.0.2
ipykernel==6.29.5
matplotlib==3.10.3
numpy==2.2.6
pandas==2.3.0
pillow==11.2.1
scikit-learn==1.7.0
timm==1.0.15
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
torchmetrics==1.7.4
transformers==4.52.4